# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Human Relay",
    "description": "Closed-alpha human capability protocol. Fake credits only.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations": {
      "post": {
        "summary": "Org",
        "operationId": "org_api_v1_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/credits": {
      "post": {
        "summary": "Credits",
        "operationId": "credits_api_v1_organization_credits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "post": {
        "summary": "Agent",
        "operationId": "agent_api_v1_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/authority": {
      "put": {
        "summary": "Authority",
        "operationId": "authority_api_v1_agents__agent_id__authority_put",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/revoke": {
      "post": {
        "summary": "Revoke",
        "operationId": "revoke_api_v1_agents__agent_id__revoke_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worker/profile": {
      "put": {
        "summary": "Profile",
        "operationId": "profile_api_v1_worker_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invitations": {
      "post": {
        "summary": "Invite",
        "operationId": "invite_api_v1_admin_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentInput": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "allowed_capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "web_research",
                "visual_verification",
                "content_judgment"
              ]
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Allowed Capabilities"
          },
          "max_task_price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Max Task Price",
            "default": "10"
          },
          "hourly_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Hourly Limit",
            "default": "20"
          },
          "daily_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Daily Limit",
            "default": "40"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Monthly Limit",
            "default": "300"
          },
          "max_concurrent_tasks": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Concurrent Tasks",
            "default": 3
          },
          "max_session_messages": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Session Messages",
            "default": 10
          },
          "max_revision_requests": {
            "type": "integer",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Revision Requests",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_name",
          "allowed_capabilities"
        ],
        "title": "AgentInput"
      },
      "CreditInput": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Amount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "CreditInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InviteInput": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "principal",
              "worker"
            ],
            "title": "Role"
          },
          "country": {
            "type": "string",
            "const": "US",
            "title": "Country"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "role",
          "country"
        ],
        "title": "InviteInput"
      },
...
      "OrganizationInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerInput": {
        "properties": {
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "web_research",
                "visual_verification",
                "content_judgment"
              ]
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Capabilities"
          },
          "blocked_categories": {
            "items": {
              "type": "string",
              "enum": [
                "web_research",
                "visual_verification",
                "content_judgment"
              ]
            },
            "type": "array",
            "maxItems": 3,
            "title": "Blocked Categories"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "busy",
              "offline"
            ],
            "title": "Availability",
            "default": "available"
          },
          "minimum_task_price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Minimum Task Price",
            "default": "3"
          },
          "maximum_active_tasks": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Maximum Active Tasks",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capabilities"
        ],
        "title": "WorkerInput"
      }
    }
  }
}